Nature made the natural numbers; All else is the work of women

+++ 0x02-functions_nested_loops/101-natural.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	unsigned int sum;
+	unsigned long int sum;
 
 	sum = 0;
